Skeleton of Landing Page has been completed
