+++ package.json
@@ -1,5 +1,7 @@
 {
   "dependencies": {
-    "bootstrap": "^5.0.2"
+    "bootstrap": "^5.0.2",
+    "react-particles-js": "^3.5.3",
+    "react-snowfall": "^1.1.0"
   }
 }
