+++ portfolio/src/App.css
@@ -42,3 +42,6 @@
   display: flex;
   justify-content: space-between;
 }
+body{
+  background-color: lightgray !important;
+}
