+++ portfolio/src/Snow.js
@@ -1,0 +1,48 @@
+import React from "react";
+import "./Home.css";
+const Snow = () => {
+
+    return (
+        <>
+            <div className="snowflakes" aria-hidden="true">
+
+                <h1>Nandhini Devi Madan</h1>
+               
+
+                <div className="snowflake">
+                    ❅
+                </div>
+                <div className="snowflake">
+                    ❅
+                </div>
+                <div className="snowflake">
+                    ❆
+                </div>
+                <div className="snowflake">
+                    ❄
+                </div>
+                <div className="snowflake">
+                    ❅
+                </div>
+                <div className="snowflake">
+                    ❆
+                </div>
+                <div className="snowflake">
+                    ❄
+                </div>
+                <div className="snowflake">
+                    ❅
+                </div>
+                <div className="snowflake">
+                    ❆
+                </div>
+                <div className="snowflake">
+                    ❄
+                </div>
+            </div>
+        </>
+
+    )
+}
+
+export default Snow;
